Fix a signed comparison and harden the DXGI adapter walk
BATTERY_UNKNOWN_RATE is 0x80000000, and bs.Rate is a signed LONG. Casting
the constant to ULONG instead of casting both sides to LONG made the
comparison mix signedness, which -Werror=sign-compare rejects. The meaning
is unchanged: that bit pattern is LONG_MIN, the value the driver writes
when it cannot measure the rate.

The adapter loop compared EnumAdapters1 against DXGI_ERROR_NOT_FOUND.
That constant exceeds LONG_MAX, so the comparison could carry the same
signedness complaint depending on how the SDK declares it. FAILED() says
what was meant, and it also removes a real defect: any error other than
NOT_FOUND used to fall through to Release() on an unset pointer.

+++ src/BatteryInfo.cpp
@@ -272,7 +272,7 @@ bool batteryinfo_bi::QueryBatteryStatus()
         info_1s.Voltage = "Unknown";
     }
 
-    info_1s.rateValid = bs.Rate != (ULONG)BATTERY_UNKNOWN_RATE;
+    info_1s.rateValid = (LONG)bs.Rate != (LONG)BATTERY_UNKNOWN_RATE;
     if (info_1s.rateValid)
     {
         info_1s.rateW = (double)(LONG)bs.Rate / 1000.0;
